+++ datastore/gcloud/aio/datastore/query_explain.py
@@ -7,8 +7,6 @@ from typing import Optional
 from typing import Tuple
 from typing import Union
 
-from .query import QueryResultBatch
-
 
 class ExplainOptions(enum.Enum):
     """Options for query explain operations."""
@@ -40,6 +38,10 @@ class IndexDefinition:
     properties: [("done", "ASC"), ("priority", "DESC"), ("__name__", "ASC")]
     """
 
+    _PROPERTIES_PATTERN = re.compile(
+        r'\s*([^\s,()]+)\s+(ASC|DESC)\s*',
+        flags=re.IGNORECASE)
+
     def __init__(self, query_scope: str = '',
                  properties: Optional[List[Tuple[str, str]]] = None):
         self.query_scope = query_scope
@@ -63,9 +65,7 @@ class IndexDefinition:
             query_scope = data['query_scope']
         if 'properties' in data:
             properties_str = data['properties']
-            properties = re.findall(r'\s*([^\s,()]+)\s+(ASC|DESC)\s*',
-                                    properties_str,
-                                    flags=re.IGNORECASE)
+            properties = cls._PROPERTIES_PATTERN.findall(properties_str)
 
         return cls(query_scope=query_scope, properties=properties)
 
@@ -208,59 +208,3 @@ class ExplainMetrics:
             explain_metrics['executionStats'] = self.execution_stats.to_repr()
 
         return explain_metrics
-
-
-class QueryResult:
-    """
-    Container class for results returned by a query operation (with or without
-    explain metrics).
-    """
-    query_result_batch_kind = QueryResultBatch
-
-    def __init__(self, result_batch: Optional[QueryResultBatch] = None,
-                 explain_metrics: Optional[ExplainMetrics] = None):
-        self.result_batch = result_batch
-        self.explain_metrics = explain_metrics
-
-    def __repr__(self) -> str:
-        return str(self.to_repr())
-
-    def __eq__(self, other: object) -> bool:
-        if not isinstance(other, QueryResult):
-            return False
-        return (self.result_batch == other.result_batch
-                and self.explain_metrics == other.explain_metrics)
-
-    @classmethod
-    def from_repr(cls, data: Dict[str, Any]) -> 'QueryResult':
-        result_batch = None
-        explain_metrics = None
-
-        if 'batch' in data:
-            result_batch = cls.query_result_batch_kind.from_repr(data['batch'])
-        if 'explainMetrics' in data:
-            explain_metrics = ExplainMetrics.from_repr(data['explainMetrics'])
-
-        return cls(result_batch=result_batch, explain_metrics=explain_metrics)
-
-    def to_repr(self) -> Dict[str, Any]:
-        result = {}
-        if self.result_batch:
-            result['batch'] = self.result_batch.to_repr()
-        if self.explain_metrics:
-            result['explainMetrics'] = self.explain_metrics.to_repr()
-
-        return result
-
-    def get_explain_metrics(self) -> Optional[ExplainMetrics]:
-        return self.explain_metrics
-
-    def get_plan_summary(self) -> Optional[PlanSummary]:
-        if self.explain_metrics is not None:
-            return self.explain_metrics.plan_summary
-        return None
-
-    def get_execution_stats(self) -> Optional[ExecutionStats]:
-        if self.explain_metrics is not None:
-            return self.explain_metrics.execution_stats
-        return None
